+++ iRate/iRate.h
@@ -1,7 +1,7 @@
 //
 //  iRate.h
 //
-//  Version 1.2.2
+//  Version 1.2.3
 //
 //  Created by Nick Lockwood on 26/01/2011.
 //  Copyright 2011 Charcoal Design. All rights reserved.
@@ -70,7 +70,7 @@
 @property (nonatomic, assign) NSUInteger appStoreID;
 
 //application name - this is set automatically
-@property (nonatomic, retain) NSString *applicationName;
+@property (nonatomic, copy) NSString *applicationName;
 
 //usage settings - these have sensible defaults
 @property (nonatomic, assign) NSUInteger usesUntilPrompt;
@@ -79,11 +79,11 @@
 @property (nonatomic, assign) float remindPeriod;
 
 //message text, you may wish to customise these, e.g. for localisation
-@property (nonatomic, retain) NSString *messageTitle;
-@property (nonatomic, retain) NSString *message;
-@property (nonatomic, retain) NSString *cancelButtonLabel;
-@property (nonatomic, retain) NSString *remindButtonLabel;
-@property (nonatomic, retain) NSString *rateButtonLabel;
+@property (nonatomic, copy) NSString *messageTitle;
+@property (nonatomic, copy) NSString *message;
+@property (nonatomic, copy) NSString *cancelButtonLabel;
+@property (nonatomic, copy) NSString *remindButtonLabel;
+@property (nonatomic, copy) NSString *rateButtonLabel;
 
 //debugging and disabling
 @property (nonatomic, assign) BOOL disabled;